+++ mc/CodeAnalysis/Syntax/Lexer.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Minsk.CodeAnalysis
+namespace Minsk.CodeAnalysis.Syntax
 {
     internal sealed class Lexer
     {
