+++ mc/CodeAnalysis/Syntax/SyntaxFacts.cs
@@ -1,4 +1,4 @@
-namespace Minsk.CodeAnalysis
+namespace Minsk.CodeAnalysis.Syntax
 {
     internal static class SyntaxFacts
     {
