+++ mc/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,4 +1,4 @@
-namespace Minsk.CodeAnalysis
+namespace Minsk.CodeAnalysis.Syntax
 {
     public enum SyntaxKind
     {
