+++ mc/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Minsk.CodeAnalysis
+namespace Minsk.CodeAnalysis.Syntax
 {
     public abstract class SyntaxNode
     {
